+++ src/main/java/org/folio/qm/validation/impl/bibliographic/BibliographicLeaderValidationRule.java
@@ -12,12 +12,17 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
+import org.folio.qm.converter.elements.LeaderItem;
 import org.folio.qm.domain.dto.MarcFormat;
 import org.folio.qm.validation.LeaderValidationRule;
 import org.folio.qm.validation.ValidationError;
 
 @Component
 public class BibliographicLeaderValidationRule implements LeaderValidationRule {
+
+  private static final List<LeaderItem> BIBLIOGRAPHIC_LEADER_ITEMS = List.of(BIB_RECORD_STATUS, BIB_RECORD_TYPE, BIBLIOGRAPHIC_LEVEL,
+    CONTROL_TYPE, CATALOGING_FORM, RESOURCE_RECORD_LEVEL);
+
   @Override
   public boolean supportFormat(MarcFormat marcFormat) {
     return MarcFormat.BIBLIOGRAPHIC == marcFormat;
@@ -25,12 +30,7 @@ public class BibliographicLeaderValidationRule implements LeaderValidationRule {
 
   @Override
   public Optional<ValidationError> validate(String leader) {
-    Optional<ValidationError> error = commonLeaderValidation(leader);
-    if (error.isPresent()) {
-      return error;
-    }
-    return validateLeaderFieldsRestrictions(leader, List.of(BIB_RECORD_STATUS, BIB_RECORD_TYPE, BIBLIOGRAPHIC_LEVEL,
-      CONTROL_TYPE, CATALOGING_FORM, RESOURCE_RECORD_LEVEL));
+    return commonLeaderValidation(leader, BIBLIOGRAPHIC_LEADER_ITEMS);
   }
 }