+++ src/test/java/org/folio/qm/validation/leader/HoldingsLeaderValidationRuleTest.java
@@ -1,7 +1,6 @@
 package org.folio.qm.validation.leader;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Optional;
@@ -9,24 +8,32 @@ import java.util.Optional;
 import org.hamcrest.core.Is;
 import org.junit.jupiter.api.Test;
 
+import org.folio.qm.domain.dto.MarcFormat;
 import org.folio.qm.validation.LeaderValidationRule;
 import org.folio.qm.validation.ValidationError;
 import org.folio.qm.validation.impl.holdings.HoldingsLeaderValidationRule;
 
 class HoldingsLeaderValidationRuleTest {
 
-  private LeaderValidationRule rule = new HoldingsLeaderValidationRule();
-  public static final String VALID_LEADER = "00241cx\\\\a2200109zn\\4500";
-  public static final String WRONG_RECORD_STATUS = "00241ex\\\\a2200109zn\\4500";
-  public static final String WRONG_RECORD_TYPE = "00241ca\\\\a2200109zn\\4500";
-  public static final String WRONG_UNDEFINED_CHARACTER = "00241cx0\\a2200109zn\\4500";
-  public static final String WRONG_ENCODING_LEVEL = "00241cx\\\\a2200109an\\4500";
-  public static final String WRONG_ITEM_INFORMATION = "00241cx\\\\a2200109za\\4500";
+  private final LeaderValidationRule rule = new HoldingsLeaderValidationRule();
+
+  private static final String VALID_LEADER = "00241cx\\\\a2200109zn\\4500";
+  private static final String WRONG_RECORD_STATUS = "00241ex\\\\a2200109zn\\4500";
+  private static final String WRONG_RECORD_TYPE = "00241ca\\\\a2200109zn\\4500";
+  private static final String WRONG_UNDEFINED_CHARACTER = "00241cx0\\a2200109zn\\4500";
+  private static final String WRONG_ENCODING_LEVEL = "00241cx\\\\a2200109an\\4500";
+  private static final String WRONG_ITEM_INFORMATION = "00241cx\\\\a2200109za\\4500";
 
 
   @Test
   void shouldValidateHoldingsLeaderWithOutErrors() {
-    assertDoesNotThrow(() -> rule.validate(VALID_LEADER));
+    var validationError = rule.validate(VALID_LEADER);
+    assertTrue(validationError.isEmpty());
+  }
+
+  @Test
+  void shouldSupportHoldingsFormat(){
+    assertTrue(rule.supportFormat(MarcFormat.HOLDINGS));
   }
 
   @Test